+++ newton/_src/viewer/__init__.py
@@ -30,6 +30,7 @@ from .viewer import ViewerBase
 from .viewer_file import ViewerFile
 from .viewer_gl import ViewerGL
 from .viewer_null import ViewerNull
+from .viewer_ovrtx import ViewerOVRTX
 from .viewer_rerun import ViewerRerun
 from .viewer_usd import ViewerUSD
 from .viewer_viser import ViewerViser
@@ -39,6 +40,7 @@ __all__ = [
     "ViewerFile",
     "ViewerGL",
     "ViewerNull",
+    "ViewerOVRTX",
     "ViewerRerun",
     "ViewerUSD",
     "ViewerViser",